reserving ticket in event service

+++ ticket-manager/event/src/main/java/pl/ticket/event/admin/ticket/repository/AdminTicketRepository.java
@@ -3,6 +3,6 @@ package pl.ticket.event.admin.ticket.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.ticket.event.admin.ticket.model.AdminTicket;
 
-public interface AdminTicketRepository extends JpaRepository<AdminTicket, Integer>
+public interface AdminTicketRepository extends JpaRepository<AdminTicket, Long>
 {
 }
